feat(conform-react): list combine command
