+++ packages/conform-dom/index.ts
@@ -1,3 +1,5 @@
+import { handleList } from './list';
+
 export type Primitive = null | undefined | string | number | boolean | Date;
 
 export type FieldElement =
@@ -46,6 +48,12 @@ export type Submission<Schema = unknown> = {
 	value: FieldValue<Schema>;
 	error: Array<[string, string]>;
 };
+
+export interface CommandButtonProps<Name extends string = string> {
+	name: `conform/${Name}`;
+	value: string;
+	formNoValidate?: boolean;
+}
 
 export function isFieldElement(element: unknown): element is FieldElement {
 	return (
@@ -230,7 +238,7 @@ export function requestSubmit(
  */
 export function requestCommand(
 	form: HTMLFormElement | undefined,
-	command: { name: string; value: string; formNoValidate?: boolean },
+	buttonProps: CommandButtonProps,
 ): void {
 	if (!form) {
 		console.warn('No form element is provided');
@@ -239,11 +247,11 @@ export function requestCommand(
 
 	const button = document.createElement('button');
 
-	button.name = command.name;
-	button.value = command.value;
+	button.name = buttonProps.name;
+	button.value = buttonProps.value;
 	button.hidden = true;
 
-	if (command.formNoValidate) {
+	if (buttonProps.formNoValidate) {
 		button.formNoValidate = true;
 	}
 
@@ -377,96 +385,4 @@ export function parse<Schema extends Record<string, any>>(
 	}
 
 	return submission as Submission<Schema>;
-}
-
-export type Command = {
-	name: string;
-	value: string;
-};
-
-export type ListCommand<Schema = unknown> =
-	| { type: 'prepend'; scope: string; payload: { defaultValue: Schema } }
-	| { type: 'append'; scope: string; payload: { defaultValue: Schema } }
-	| {
-			type: 'replace';
-			scope: string;
-			payload: { defaultValue: Schema; index: number };
-	  }
-	| { type: 'remove'; scope: string; payload: { index: number } }
-	| { type: 'reorder'; scope: string; payload: { from: number; to: number } };
-
-export function parseListCommand<Schema = unknown>(
-	data: string,
-): ListCommand<Schema> {
-	try {
-		const command = JSON.parse(data);
-
-		if (
-			typeof command.type !== 'string' ||
-			!['prepend', 'append', 'replace', 'remove', 'reorder'].includes(
-				command.type,
-			)
-		) {
-			throw new Error('Unsupported list command type');
-		}
-
-		return command;
-	} catch (error) {
-		throw new Error(`Invalid list command: "${data}"; ${error}`);
-	}
-}
-
-export function updateList<Schema>(
-	list: Array<Schema>,
-	command: ListCommand<Schema>,
-): Array<Schema> {
-	switch (command.type) {
-		case 'prepend': {
-			list.unshift(command.payload.defaultValue);
-			break;
-		}
-		case 'append': {
-			list.push(command.payload.defaultValue);
-			break;
-		}
-		case 'replace': {
-			list.splice(command.payload.index, 1, command.payload.defaultValue);
-			break;
-		}
-		case 'remove':
-			list.splice(command.payload.index, 1);
-			break;
-		case 'reorder':
-			list.splice(
-				command.payload.to,
-				0,
-				...list.splice(command.payload.from, 1),
-			);
-			break;
-		default:
-			throw new Error('Unknown list command received');
-	}
-
-	return list;
-}
-
-export function handleList<Schema>(
-	submission: Submission<Schema>,
-): Submission<Schema> {
-	if (submission.type !== 'list') {
-		return submission;
-	}
-
-	const command = parseListCommand(submission.intent ?? '');
-	const paths = getPaths(command.scope);
-
-	setValue(submission.value, paths, (list) => {
-		if (!Array.isArray(list)) {
-			throw new Error('The list command can only be applied to a list');
-		}
-
-		return updateList(list, command);
-	});
-
-	return submission;
 }
